finished with TBB download

+++ setup.py
@@ -6,23 +6,13 @@ import shutil
 import urllib
 import tempfile
 import zipfile
+import tarfile
 from setuptools.command.build_ext import build_ext
 from setuptools.command.develop import develop
-from distutils.core import setup, Command
+from distutils.dep_util import newer_group
 import numpy as np
 
 ''' C extensions
-
-PETSc Linking
--------------
-
-If you want to setup SimNIBS to use your own locally compiled PETSC, you can call
-setup.py with the SIMNIBS_PETSC_INCLUDE and SIMNIBS_PETSC_DIR environment variables, for
-example
-
-export SIMNIBS_PETSC_DIR="/path/to/petsc/simnibs_petsc_arch/lib"
-export SIMNIBS_PETSC_INCLUDE="/path/to/petsc/simnibs_direct_petsc_arch/include"
-python setup.py develop
 
 CGAL Compilation
 -----------------
@@ -38,13 +28,13 @@ Boost can be instaled with
     Ubuntu: sudo apt install libboost-all-dev
     MacOS: brew install boost
     Windows: conda install boost
-
     Boost is also header-only, so we only need it during compile time
 
 For more info, refer to https://doc.cgal.org/latest/Manual/thirdparty.html
 
 '''
 
+# Information for CGAL download
 CGAL_version = '5.0'  # I tried 5.0.1 but tests fail!
 CGAL_headers = os.path.abspath(f'CGAL-{CGAL_version}/include')
 CGAL_url = (
@@ -53,6 +43,7 @@ CGAL_url = (
     f'CGAL-{CGAL_version}-library.zip'
 )
 
+# Information for eigen download
 eigen_version = '3.3.7'
 eigen_headers = os.path.abspath(f'eigen-{eigen_version}')
 eigen_url = (
@@ -60,29 +51,58 @@ eigen_url = (
     f'archive/{eigen_version}/eigen-{eigen_version}.zip'
 )
 
+# Information for Intel TBB download
 tbb_version = '2020.1'
-tbb_os = {'win32': 'win', 'linux': 'lin', 'darwin': 'mac'}[sys.platform]
-tbb_extension = 'zip' if sys.platform == 'win32' else 'tgz'
-tbb_headers = os.path.abspath(f'tbb-{tbb_version}-{tbb_os}/tbb/include/')
-tbb_url = (
-    f'https://github.com/intel/tbb/releases/download/'
-    f'v{tbb_version}/tbb-{tbb_version}-{tbb_os}.{tbb_extension}'
-)
+tbb_path = os.path.abspath('tbb')
+tbb_headers = os.path.join(tbb_path, 'tbb', 'include')
+if sys.platform == 'win32':
+    tbb_url = (
+        f'https://github.com/intel/tbb/releases/download/'
+        f'v{tbb_version}/tbb-{tbb_version}-win.zip'
+    )
+    tbb_libs = [
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'vc14_uwp', 'tbb.def'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'vc14_uwp', 'tbb.lib'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'vc14_uwp', 'tbbmalloc.lib'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'vc14_uwp', 'tbbmalloc.def'),
+    ]
+elif sys.platform == 'linux':
+    tbb_url = (
+        f'https://github.com/intel/tbb/releases/download/'
+        f'v{tbb_version}/tbb-{tbb_version}-lin.tgz'
+    )
+    tbb_libs = [
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'gcc4.8', 'libtbb.so'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'gcc4.8', 'libtbb.so.2'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'gcc4.8', 'libtbbmalloc.so'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'intel64', 'gcc4.8', 'libtbbmalloc.so.2'),
+    ]
+elif sys.platform == 'darwin':
+    tbb_url = (
+        f'https://github.com/intel/tbb/releases/download/'
+        f'v{tbb_version}/tbb-{tbb_version}-mac.tgz'
+    )
+    tbb_libs = [
+        os.path.join(tbb_path, 'tbb', 'lib', 'libtbb.dylib'),
+        os.path.join(tbb_path, 'tbb', 'lib', 'libtbbmalloc.dylib'),
+    ]
+else:
+    raise OSError('OS not supported!')
+
+
+# Setup compilation
 is_conda = 'CONDA_PREFIX' in os.environ
 
+# Windows compilation
 if sys.platform == 'win32':
     petsc_libs = ['libpetsc', 'msmpi']
-    if 'SIMNIBS_PETSC_INCLUDE' in os.environ:
-        petsc_include = [np.get_include()] + os.environ['SIMNIBS_PETSC_INCLUDE'].split(';')
-    else:
-        petsc_include = [np.get_include(),
-                         'simnibs/external/include/win/petsc',
-                         'simnibs/external/include/win/hypre',
-                         'simnibs/external/include/win/mpi']
-    if 'SIMNIBS_PETSC_DIR' in os.environ:
-        petsc_dirs = [os.environ['SIMNIBS_PETSC_DIR']]
-    else:
-        petsc_dirs = ['simnibs/external/lib/win']
+    petsc_include = [
+        np.get_include(),
+        'simnibs/external/include/win/petsc',
+        'simnibs/external/include/win/hypre',
+        'simnibs/external/include/win/mpi'
+    ]
+    petsc_dirs = ['simnibs/external/lib/win']
     petsc_runtime = None
 
     cgal_libs = ['libmpfr-4', 'libgmp-10', 'zlib']
@@ -90,33 +110,32 @@ if sys.platform == 'win32':
         np.get_include(),
         CGAL_headers,
         eigen_headers,
+        tbb_headers,
         'simnibs/external/include/win/mpfr',
         'simnibs/external/include/win/gmp'
     ]
     if is_conda:
         cgal_include += [os.path.join(os.environ['CONDA_PREFIX'], 'Library', 'include')]
     cgal_dirs = ['simnibs/external/lib/win']
+    #TODO
     cgal_runtime = None
     cgal_compile_args = None
 
 elif sys.platform == 'linux':
     petsc_libs = ['petsc']
-    if 'SIMNIBS_PETSC_INCLUDE' in os.environ:
-        petsc_include = [np.get_include()] + os.environ['SIMNIBS_PETSC_INCLUDE'].split(':')
-    else:
-        petsc_include = [np.get_include(), 'simnibs/external/include/linux/petsc']
-    if 'SIMNIBS_PETSC_DIR' in os.environ:
-        petsc_dirs = [os.environ['SIMNIBS_PETSC_DIR']]
-        petsc_runtime = petsc_dirs
-    else:
-        petsc_dirs = ['simnibs/external/lib/linux']
-        petsc_runtime = ['$ORIGIN/../external/lib/linux']
+    petsc_include = [
+        np.get_include(),
+        'simnibs/external/include/linux/petsc'
+    ]
+    petsc_dirs = ['simnibs/external/lib/linux']
+    petsc_runtime = ['$ORIGIN/../external/lib/linux']
 
-    cgal_libs = ['mpfr', 'gmp', 'z']
+    cgal_libs = ['mpfr', 'gmp', 'z', 'tbb', 'tbbmalloc', 'pthread']
     cgal_include = [
         np.get_include(),
         CGAL_headers,
         eigen_headers,
+        tbb_headers,
         'simnibs/external/include/linux/mpfr',
         'simnibs/external/include/linux/gmp'
     ]
@@ -124,7 +143,6 @@ elif sys.platform == 'linux':
         cgal_include += [os.path.join(os.environ['CONDA_PREFIX'], 'include')]
     cgal_dirs = ['simnibs/external/lib/linux']
     cgal_runtime = ['$ORIGIN/../external/lib/linux']
-    # TODO: setup compile args in other platforms
     cgal_compile_args = [
         '-Os', '-flto',
         '-DCGAL_CONCURRENT_MESH_3',
@@ -133,34 +151,28 @@ elif sys.platform == 'linux':
         '-DNOMINMAX',
         '-DCGAL_EIGEN3_ENABLED',
         '-DCGAL_LINKED_WITH_TBB',
+        '-DNDEBUG',
+        '-DCGAL_USE_ZLIB=1',
         '-frounding-math',
         '-std=gnu++14',
-
     ]
-    # TODO: Add libtbb
-    cgal_link_args = [
-        '/usr/lib/x86_64-linux-gnu/libtbb.so',
-        '/usr/lib/x86_64-linux-gnu/libtbbmalloc.so',
-        '-lpthread'
-    ]
+    cgal_link_args = None
 
 elif sys.platform == 'darwin':
     petsc_libs = ['petsc']
-    if 'SIMNIBS_PETSC_INCLUDE' in os.environ:
-        petsc_include = [np.get_include()] + os.environ['SIMNIBS_PETSC_INCLUDE'].split(':')
-    else:
-        petsc_include = [np.get_include(), 'simnibs/external/include/osx/petsc']
-    if 'SIMNIBS_PETSC_DIR' in os.environ:
-        petsc_dirs = [os.environ['SIMNIBS_PETSC_DIR']]
-    else:
-        petsc_dirs = ['simnibs/external/lib/osx']
+    petsc_include = [
+        np.get_include(),
+        'simnibs/external/include/osx/petsc'
+    ]
+    petsc_dirs = ['simnibs/external/lib/osx']
     petsc_runtime = None
 
-    cgal_libs = ['mpfr', 'gmp', 'z']
+    cgal_libs = ['mpfr', 'gmp', 'z', 'tbb', 'tbbmalloc']
     cgal_include = [
         np.get_include(),
         CGAL_headers,
         eigen_headers,
+        tbb_headers,
         'simnibs/external/include/osx/mpfr',
         'simnibs/external/include/osx/gmp'
     ]
@@ -187,40 +199,54 @@ elif sys.platform == 'darwin':
         '-stdlib=libc++'
     ]
 
-
 else:
     raise OSError('OS not supported!')
 
-extension = [
-    Extension('simnibs.mesh_tools.cython_msh',
-              ["simnibs/mesh_tools/cython_msh.pyx"],
-              include_dirs=[np.get_include()]),
-    Extension('simnibs.mesh_tools._marching_cubes_lewiner_cy',
-              ["simnibs/mesh_tools/_marching_cubes_lewiner_cy.pyx"],
-              include_dirs=[np.get_include()]),
-    Extension('simnibs.segmentation._cs_utils',
-              ["simnibs/segmentation/_cs_utils.pyx",
-               "simnibs/external/cat12/genus0.c"],
-              include_dirs=[np.get_include(),
-                            'simnibs/external/cat12']),
-    Extension('simnibs._compiled.petsc_solver',
-              ["simnibs/_compiled/petsc_solver.pyx"],
-              include_dirs=petsc_include,
-              library_dirs=petsc_dirs,
-              libraries=petsc_libs,
-              runtime_library_dirs=petsc_runtime),
-    Extension('simnibs._compiled.create_mesh',
-              sources=["simnibs/_compiled/create_mesh.pyx"],
-              language='c++',
-              include_dirs=cgal_include,
-              libraries=cgal_libs,
-              library_dirs=cgal_dirs,
-              runtime_library_dirs=cgal_runtime,
-              extra_compile_args=cgal_compile_args,
-              extra_link_args=cgal_link_args),
+cython_msh = Extension(
+    'simnibs.mesh_tools.cython_msh',
+    ["simnibs/mesh_tools/cython_msh.pyx"],
+    include_dirs=[np.get_include()]
+)
+marching_cubes_lewiner_cy = Extension(
+    'simnibs.mesh_tools._marching_cubes_lewiner_cy',
+    ["simnibs/mesh_tools/_marching_cubes_lewiner_cy.pyx"],
+    include_dirs=[np.get_include()]
+)
+cs_utils = Extension(
+    'simnibs.segmentation._cs_utils',
+    ["simnibs/segmentation/_cs_utils.pyx", "simnibs/external/cat12/genus0.c"],
+    include_dirs=[np.get_include(), 'simnibs/external/cat12']
+)
+petsc_solver = Extension(
+    'simnibs._compiled.petsc_solver',
+    sources=["simnibs/_compiled/petsc_solver.pyx"],
+    depends=["simnibs/_compiled/_solver.c"],
+    include_dirs=petsc_include,
+    library_dirs=petsc_dirs,
+    libraries=petsc_libs,
+    runtime_library_dirs=petsc_runtime
+)
+create_mesh = Extension(
+    'simnibs._compiled.create_mesh',
+    sources=["simnibs/_compiled/create_mesh.pyx"],
+    depends=["simnibs/_compiled/_mesh.cpp"],
+    language='c++',
+    include_dirs=cgal_include,
+    libraries=cgal_libs,
+    library_dirs=cgal_dirs,
+    runtime_library_dirs=cgal_runtime,
+    extra_compile_args=cgal_compile_args,
+    extra_link_args=cgal_link_args
+)
+
+
+extensions = [
+    cython_msh,
+    marching_cubes_lewiner_cy,
+    cs_utils,
+    petsc_solver,
+    create_mesh
 ]
-
-
 ####################################################
 # Add all scripts in the cli folder and a bit more #
 ####################################################
@@ -245,18 +271,45 @@ gui_scripts = [
 ]
 
 
-def download_and_extract(url):
+def download_and_extract(url, path='.'):
+    ''' Downloads and extracts a zip or tar-gz folder '''
     with urllib.request.urlopen(url) as response:
         with tempfile.NamedTemporaryFile('wb', delete=False) as tmpf:
             shutil.copyfileobj(response, tmpf)
             tmpname = tmpf.name
 
-    with zipfile.ZipFile(tmpname) as z:
-        z.extractall()
+    if url.endswith('.zip'):
+        with zipfile.ZipFile(tmpname) as z:
+            z.extractall(path)
+
+    elif url.endswith('.tgz') or url.endswith('.tar.gz'):
+        with tarfile.open(tmpname, 'r:gz') as z:
+            z.extractall(path)
+    else:
+        raise IOError('Could not extract file, unrecognized extension')
 
     os.remove(tmpname)
 
+
+def install_lib(url, path, libs):
+    ''' Downloads a compiled library from the internet and move to "lib" folder '''
+    download_and_extract(url, path)
+    if sys.platform == 'darwin':
+        folder_name = 'osx'
+    elif sys.platform == 'linux':
+        folder_name = 'linux'
+    elif sys.platform == 'win32':
+        folder_name = 'win'
+    for l in libs:
+        shutil.copy(
+            l, f'simnibs/external/lib/{folder_name}',
+            follow_symlinks=False
+        )
+
 def move_libraries(build_folder, operation=shutil.move):
+    '''
+        Move librariessimnibs/external/lib -> simnibs/_compiled
+    '''
     if sys.platform == 'darwin':
         folder_name = 'osx'
     elif sys.platform == 'linux':
@@ -272,36 +325,47 @@ def move_libraries(build_folder, operation=shutil.move):
             os.path.join(compliled_folder, os.path.basename(fn))
         )
 
+
 class build_ext_(build_ext):
+    '''
+        Build the extension, download some dependencies and remove stuff from other OS
+    '''
     def run(self):
         from Cython.Build import cythonize
+        ## Cythonize
         self.extension = cythonize(self.extensions)
-        download_and_extract(CGAL_url)
-        download_and_extract(eigen_url)
+        ## Download requirements
+        changed_meshing = newer_group(
+            create_mesh.sources + create_mesh.depends,
+            self.get_ext_fullpath(create_mesh.name),
+            'newer'
+        )
+        if self.force or changed_meshing:
+            download_and_extract(CGAL_url)
+            download_and_extract(eigen_url)
+            install_lib(tbb_url, tbb_path, tbb_libs)
+        # Compile
         build_ext.run(self)
+        # cleanup downloads
+        if self.force or changed_meshing:
+            shutil.rmtree(f'CGAL-{CGAL_version}', ignore_errors=True)
+            shutil.rmtree(f'eigen-{eigen_version}', ignore_errors=True)
+            shutil.rmtree(tbb_path, ignore_errors=True)
         # Remove unescessary binary files
-        shutil.rmtree(f'CGAL-{CGAL_version}', ignore_errors=True)
-        shutil.rmtree(f'eigen-{eigen_version}', ignore_errors=True)
         linux_folders = [
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'bin', 'linux'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'include', 'linux'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'lib', 'linux'),
-            os.path.join(self.build_lib, 'simnibs', 'resources',
-                         'spm12', 'toolbox', 'cat12', 'CAT.glnx86'),
         ]
         osx_folders = [
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'bin', 'osx'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'include', 'osx'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'lib', 'osx'),
-            os.path.join(self.build_lib, 'simnibs', 'resources',
-                         'spm12', 'toolbox', 'cat12', 'CAT.maci64'),
         ]
         win_folders = [
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'bin', 'win'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'include', 'win'),
             os.path.join(self.build_lib, 'simnibs', 'extenal', 'lib', 'win'),
-            os.path.join(self.build_lib, 'simnibs', 'resources',
-                         'spm12', 'toolbox', 'cat12', 'CAT.w32'),
         ]
         if sys.platform == 'linux':
             [shutil.rmtree(f, True) for f in osx_folders]
@@ -326,6 +390,7 @@ class develop_(develop):
         if sys.platform == 'darwin':
             move_libraries('.', shutil.copy)
 
+
 setup(name='simnibs',
       version=open("simnibs/_version.py").readlines()[-1].split()[-1].strip("\"'"),
       description='www.simnibs.org',
@@ -333,7 +398,7 @@ setup(name='simnibs',
       author_email='support@simnibs.org',
       packages=find_packages(),
       license='GPL3',
-      ext_modules=extension,
+      ext_modules=extensions,
       include_package_data=True,
       cmdclass={
           'build_ext': build_ext_,
